requirements.txt
aiogram==2.9.2
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.2.0
Babel==2.14.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
colorama==0.4.6
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
loguru==0.7.2
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector==2.2.9
psycopg==3.1.10
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.0.3
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.25
typing-extensions==4.9.0
tzdata==2023.4
win32-setctime==1.1.0
yarl==1.9.4
zipp==3.17.0
